+++ packages/stream_chat/CHANGELOG.md
@@ -1,3 +1,14 @@
+## Upcoming
+
+✅ Added
+
+- Added `Filter.notExists`.
+
+🐞 Fixed
+
+- [[#710]](https://github.com/GetStream/stream-chat-flutter/issues/710) Fixed JWT requiring using `String` as id.
+- Fixed expired CDN attachment links not updating correctly.
+
 ## 3.0.0
 
 🛑️ Breaking Changes from `2.2.1`